Split text_trust and improved documentation
